chore(tests): consolidate e2e test location
