+++ test/e2e/README.md
@@ -6,6 +6,7 @@ This directory contains end-to-end tests for the Phoenix (SA-OMF) project.
 
 - `/test/e2e/benchmarks/` - End-to-end benchmarks for measuring performance
 - `/test/e2e/integration/` - Integration tests for verifying component interactions
+- All end-to-end tests reside under `/test/e2e/`. The former `/test/e2e_tests/` directory has been removed.
 
 ## Running the Tests
